🪒 Razor: Flatten modules and remove single-variant enums
This commit enforces the KISS principle by identifying and simplifying over-engineered areas of the codebase:

1. **Flattened Deep Hierarchies:**
   - Moved `relvar-core/src/query/mod.rs` to `relvar-core/src/query.rs` and deleted the `query/` directory.
   - Merged `relvar-core/src/storage_engine/in_memory.rs` into `relvar-core/src/storage_engine/mod.rs` and renamed it to `storage_engine.rs`, deleting the `storage_engine/` directory. These deep directories contained very low file counts (1 and 2 files, respectively), violating simplicity rules. The larger `mvcc` directory (5 files, >1k LOC) was left intact.

2. **Removed Single-Variant Enums:**
   - Replaced `UnionError`, `IntersectError`, and `DifferenceError` single-variant `TypeMismatch` enums with `pub struct [Name]Error;` unit structs. This removes unnecessary indirection and pattern-matching boilerplate from both the library and test code.

Verified via `cargo clippy` and `cargo test`. All temporary scripts have been cleaned up.
