+++ relvar-core/src/algebra/difference.rs
@@ -38,14 +38,9 @@ use thiserror::Error;
 
 /// Errors that can occur during difference operations.
 #[derive(Debug, Error)]
-pub enum DifferenceError {
-    /// The two relations have incompatible types (different headings).
-    ///
-    /// Difference requires both relations to have exactly the same heading
-    /// (attribute names and types).
-    #[error("Relations must have the same type (heading) for difference")]
-    TypeMismatch,
-}
+#[error("Relations must have the same type (heading) for difference")]
+pub struct DifferenceError;
+
 
 impl Relation {
     /// Computes the set difference of this relation with another.
@@ -65,7 +60,7 @@ impl Relation {
     ///
     /// # Errors
     ///
-    /// Returns [`DifferenceError::TypeMismatch`] if the relations have different
+    /// Returns [`DifferenceError`] if the relations have different
     /// headings (different attribute names or types).
     ///
     /// # Behavior
@@ -108,7 +103,7 @@ impl Relation {
     pub fn difference(&self, other: &Relation) -> Result<Self, DifferenceError> {
         // Check type compatibility
         if self.relation_type() != other.relation_type() {
-            return Err(DifferenceError::TypeMismatch);
+            return Err(DifferenceError);
         }
 
         // Filter tuples and create relation without redundant checks
@@ -206,7 +201,7 @@ mod tests {
 
         let result = rel1.difference(&rel2);
         assert!(result.is_err());
-        assert!(matches!(result.unwrap_err(), DifferenceError::TypeMismatch));
+        assert!(matches!(result.unwrap_err(), DifferenceError));
     }
 
     #[test]
