+++ relvar-core/src/algebra/intersect.rs
@@ -39,14 +39,9 @@ use thiserror::Error;
 
 /// Errors that can occur during intersection operations.
 #[derive(Debug, Error)]
-pub enum IntersectError {
-    /// The two relations have incompatible types (different headings).
-    ///
-    /// Intersection requires both relations to have exactly the same heading
-    /// (attribute names and types).
-    #[error("Relations must have the same type (heading) for intersection")]
-    TypeMismatch,
-}
+#[error("Relations must have the same type (heading) for intersection")]
+pub struct IntersectError;
+
 
 impl Relation {
     /// Computes the intersection of this relation with another.
@@ -66,7 +61,7 @@ impl Relation {
     ///
     /// # Errors
     ///
-    /// Returns [`IntersectError::TypeMismatch`] if the relations have different
+    /// Returns [`IntersectError`] if the relations have different
     /// headings (different attribute names or types).
     ///
     /// # Behavior
@@ -109,7 +104,7 @@ impl Relation {
     pub fn intersect(&self, other: &Relation) -> Result<Self, IntersectError> {
         // Check type compatibility
         if self.relation_type() != other.relation_type() {
-            return Err(IntersectError::TypeMismatch);
+            return Err(IntersectError);
         }
 
         // Filter tuples and create relation without redundant checks
@@ -177,7 +172,7 @@ mod tests {
 
         let result = rel1.intersect(&rel2);
         assert!(result.is_err());
-        assert!(matches!(result.unwrap_err(), IntersectError::TypeMismatch));
+        assert!(matches!(result.unwrap_err(), IntersectError));
     }
 
     #[test]
